+++ test/YamlMessageSourceTest.php
@@ -2,7 +2,6 @@
 declare(strict_types=1);
 namespace yii\i18n;
 
-use function PHPUnit\Expect\{expect, it};
 use PHPUnit\Framework\{TestCase};
 
 /**
@@ -11,18 +10,18 @@ use PHPUnit\Framework\{TestCase};
 class YamlMessageSourceTest extends TestCase {
 
   /**
-   * @test YamlMessageSource::flatten
+   * Tests the `YamlMessageSource::flatten
    */
-  public function testFlatten(): void {
+  function testFlatten(): void {
     $flatten = function($array) {
       return $this->flatten($array);
     };
 
-    it('should merge the keys of a multidimensional array', function() use ($flatten) {
+    // It should merge the keys of a multidimensional array.
       $model = new YamlMessageSource;
-      expect($flatten->call($model, []))->to->equal([]);
-      expect($flatten->call($model, ['foo' => 'bar', 'baz' => 'qux']))->to->equal(['foo' => 'bar', 'baz' => 'qux']);
-      expect($flatten->call($model, ['foo' => ['bar' => 'baz']]))->to->equal(['foo.bar' => 'baz']);
+      assertThat($flatten->call($model, []), equalTo([]);
+      assertThat($flatten->call($model, ['foo' => 'bar', 'baz' => 'qux']), equalTo(['foo' => 'bar', 'baz' => 'qux']);
+      assertThat($flatten->call($model, ['foo' => ['bar' => 'baz']]), equalTo(['foo.bar' => 'baz']);
 
       $source = [
         'foo' => 'bar',
@@ -33,7 +32,7 @@ class YamlMessageSourceTest extends TestCase {
         ]]
       ];
 
-      expect($flatten->call($model, $source))->to->equal([
+      assertThat($flatten->call($model, $source), equalTo([
         'foo' => 'bar',
         'bar.baz' => 'qux',
         'baz.qux.foo' => 'bar',
@@ -41,7 +40,7 @@ class YamlMessageSourceTest extends TestCase {
       ]);
     });
 
-    it('should allow different nesting separators', function() use ($flatten) {
+    // It should allow different nesting separators.
       $source = [
         'foo' => 'bar',
         'bar' => ['baz' => 'qux'],
@@ -52,7 +51,7 @@ class YamlMessageSourceTest extends TestCase {
       ];
 
       $model = new YamlMessageSource(['nestingSeparator' => '/']);
-      expect($flatten->call($model, $source))->to->equal([
+      assertThat($flatten->call($model, $source), equalTo([
         'foo' => 'bar',
         'bar/baz' => 'qux',
         'baz/qux/foo' => 'bar',
@@ -60,7 +59,7 @@ class YamlMessageSourceTest extends TestCase {
       ]);
 
       $model = new YamlMessageSource(['nestingSeparator' => '->']);
-      expect($flatten->call($model, $source))->to->equal([
+      assertThat($flatten->call($model, $source), equalTo([
         'foo' => 'bar',
         'bar->baz' => 'qux',
         'baz->qux->foo' => 'bar',
@@ -70,73 +69,73 @@ class YamlMessageSourceTest extends TestCase {
   }
 
   /**
-   * @test YamlMessageSource::getMessageFilePath
+   * Tests the `YamlMessageSource::getMessageFilePath
    */
-  public function testGetMessageFilePath(): void {
+  function testGetMessageFilePath(): void {
     $getMessageFilePath = function($category, $language) {
       return $this->getMessageFilePath($category, $language);
     };
 
-    it('should return the proper path to the message file', function() use ($getMessageFilePath) {
+    // It should return the proper path to the message file.
       $model = new YamlMessageSource(['basePath' => '@root/test/fixtures']);
       $messageFile = str_replace('/', DIRECTORY_SEPARATOR, __DIR__.'/fixtures/fr/messages.yaml');
-      expect($getMessageFilePath->call($model, 'messages', 'fr'))->to->equal($messageFile);
+      assertThat($getMessageFilePath->call($model, 'messages', 'fr'), equalTo($messageFile);
     });
 
-    it('should should support different file extensions', function() use ($getMessageFilePath) {
+    // It should should support different file extensions.
       $model = new YamlMessageSource(['basePath' => '@root/test/fixtures', 'fileExtension' => 'yml']);
       $messageFile = str_replace('/', DIRECTORY_SEPARATOR, __DIR__.'/fixtures/fr/messages');
-      expect($getMessageFilePath->call($model, 'messages', 'fr'))->to->equal("$messageFile.yml");
+      assertThat($getMessageFilePath->call($model, 'messages', 'fr'), equalTo("$messageFile.yml");
     });
   }
 
   /**
-   * @test YamlMessageSource::loadMessagesFromFile
+   * Tests the `YamlMessageSource::loadMessagesFromFile
    */
-  public function testLoadMessagesFromFile(): void {
+  function testLoadMessagesFromFile(): void {
     $loadMessagesFromFile = function($messageFile) {
       return $this->loadMessagesFromFile($messageFile);
     };
 
-    it('should properly load the JSON source and parse it as array', function() use ($loadMessagesFromFile) {
+    // It should properly load the JSON source and parse it as array.
       $model = new YamlMessageSource(['basePath' => '@root/test/fixtures', 'enableNesting' => true]);
       $messageFile = \Yii::getAlias("{$model->basePath}/fr/messages.yaml");
-      expect($loadMessagesFromFile->call($model, $messageFile))->to->equal([
+      assertThat($loadMessagesFromFile->call($model, $messageFile), equalTo([
         'Hello World!' => 'Bonjour le monde !',
         'foo.bar.baz' => 'FooBarBaz'
       ]);
     });
 
-    it('should enable proper translation of source strings', function() {
+    // It should enable proper translation of source strings.
       $model = new YamlMessageSource(['basePath' => '@root/test/fixtures', 'enableNesting' => true]);
-      expect($model->translate('messages', 'Hello World!', 'fr'), 'Bonjour le monde !');
-      expect($model->translate('messages', 'foo.bar.baz', 'fr'), 'FooBarBaz');
+      assertThat($model->translate('messages', 'Hello World!', 'fr'), 'Bonjour le monde !');
+      assertThat($model->translate('messages', 'foo.bar.baz', 'fr'), 'FooBarBaz');
 
       $model->nestingSeparator = '/';
-      expect($model->translate('messages', 'foo/bar/baz', 'fr'), 'FooBarBaz');
+      assertThat($model->translate('messages', 'foo/bar/baz', 'fr'), 'FooBarBaz');
     });
   }
 
   /**
-   * @test YamlMessageSource::parseMessages
+   * Tests the `YamlMessageSource::parseMessages
    */
-  public function testParseMessages(): void {
+  function testParseMessages(): void {
     $parseMessages = function($messageData) {
       return $this->parseMessages($messageData);
     };
 
-    it('should parse a YAML file as a hierarchical array', function() use ($parseMessages) {
+    // It should parse a YAML file as a hierarchical array.
       $model = new YamlMessageSource(['basePath' => '@root/test/fixtures', 'enableNesting' => true]);
       $messages = $parseMessages->call($model, file_get_contents(\Yii::getAlias("{$model->basePath}/fr/messages.yaml")));
-      expect($messages)->equal([
+      assertThat($messages)->equal([
         'Hello World!' => 'Bonjour le monde !',
         'foo' => ['bar' => ['baz' => 'FooBarBaz']]
       ]);
     });
 
-    it('should parse an invalid YAML file as an empty array', function() use ($parseMessages) {
+    // It should parse an invalid YAML file as an empty array.
       $model = new YamlMessageSource(['basePath' => '@root/test/fixtures']);
-      expect($parseMessages->call($model, ''))->to->be->empty;
+      assertThat($parseMessages->call($model, ''), isEmpty());
     });
   }
 }