+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Breaking change: renamed the package to `yii-i18n-messages`.
 - Added the `FileMessageSource` abstract class.
 - Added the `ExtendedPhpMessageSource` and `YamlMessageSource` classes.
+- Dropped the dependency on [PHPUnit-Expect](https://dev.belin.io/phpunit-expect).
 
 ## Version [5.0.0](https://github.com/cedx/yii2-i18n-messages/compare/v4.1.0...v5.0.0)
 - Breaking change: raised the required [PHP](https://secure.php.net) version.
@@ -99,7 +100,7 @@
 - Lowered the required [PHP](https://secure.php.net) version.
 
 ## Version [0.1.1](https://github.com/cedx/yii2-i18n-messages/compare/v0.1.0...v0.1.1)
-- Using [Gulp.js](http://gulpjs.com) as build system.
+- Using [Gulp](https://gulpjs.com) as build system.
 
 ## Version 0.1.0
 - Initial release.